incident-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.4.0
pymongo==4.6.3
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
PyJWT==2.8.0